+++ src/CalendarBundle/Resources/public/css/calendar_main.css
@@ -1,48 +1,3 @@
-/************************************************
-*
-*   1. Taille du Calendar
-$
-************************************************/
-
-body {
-    padding: 0;
-    font-family: "Lucida Grande",Helvetica,Arial,Verdana,sans-serif;
-    font-size: 14px;
-}
-
-#calendar {
-    max-width: 900px;
-    margin: 0 auto;
-}
-
-/************************************************
-*
-*   2. Boite modal
-$
-************************************************/
-
-.modal-corps{
-    background-color: #004d40;
-    color: #dff1ff;
-}
-
-.modal-header{
-
-}
-
-.modal-body{
-
-}
-
-.modal-body-image{
-    background-color: #9F9F9F;
-    color: #3b3b1f;
-}
-
-.modal-footer{
-    background-color: #9F9F9F;
-}
-
 /*********************************************************/
 /*********************************************************/
 /*********************** PAGE NEW ************************/
@@ -51,7 +6,7 @@ $
 
 /************************************************
 *
-*   3. Color Picker
+*   1. Color Picker
 $
 ************************************************/
