+++ src/CalendarBundle/Resources/public/js/calendar/fullcalendar.js
