+++ src/CalendarBundle/Resources/views/agenda/edit.html.twig
@@ -4,9 +4,6 @@
 
     {{ form_start(edit_form) }}
 
-{#<div id="modalEdit" class="modal bottom-sheet">#}
-    {#<div class="modal-content">#}
-
     <h1>Agenda edit</h1>
 
     <div class="container">
@@ -55,6 +52,8 @@
                     {{ form_errors(edit_form.color) }}
             </div>
 
+{#--------------------------------------------------------------------------------------------------------------------#}
+{#-------------------------------- RESERVE POUR PLUS TARD QUAND LE SLIDE SERA ACTIF ----------------------------------#}
 {#--------------------------------------------------------------------------------------------------------------------#}
             {#<div class="col s12 m6 l6">#}
                 {#<div class="checkbox">#}
