+++ src/CalendarBundle/Resources/views/layout.html.twig
@@ -8,6 +8,11 @@
 
     {% block javascripts %}
         {{ parent() }}
-        <script src="{{ asset('bundles/calendar/js/main_calendar.js') }}"></script>
         <script src="{{ asset('bundles/calendar/js/materialize-colorpicker.min.js') }}"></script>
+        <script>
+            $('#colorpicker').colorpicker({
+                format: 'hex',
+                color: 'transparent'
+            });
+        </script>
     {% endblock %}
