+++ src/PlateFormeBundle/Resources/views/Default/basePlateforme.html.twig
@@ -5,6 +5,8 @@
         <title>{% block title %}PLateforme du Haras de la métamorphose{% endblock %}</title>
         <link rel="icon" type="image/x-icon" href="{{ asset('bundles/plateforme/images/icone-cheval.png') }}" />
         <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/materialize/0.97.8/css/materialize.min.css">
+        <link rel="stylesheet" href="{{ asset('bundles/calendar/css/fullcalendar.css') }}">
+        <link rel="stylesheet" href="{{ asset('bundles/calendar/css/modal.css') }}">
         <link rel="stylesheet" href="{{ asset('bundles/haras/styles/stylesheets/style.css') }}">
 
         <!-- Icons -->
