+++ src/PlateFormeBundle/Resources/views/Default/index.html.twig
@@ -4,5 +4,4 @@
 
 
 
-
 {% endblock %}
