+++ src/PlateFormeBundle/Resources/views/layout.html.twig
@@ -74,10 +74,6 @@
     </header>
 {% endblock %}
 
-{% block body %}
-{% endblock %}
-
-
 {% block footer %}
     <footer class="page-footer">
         <div class="container">
@@ -106,14 +102,4 @@
             </div>
         </div>
     </footer>
-{% endblock %}
-
-{% block javascripts %}
-
-    {{ parent() }}
-
-    <script>
-
-    </script>
-
 {% endblock %}
